requirements.txt
opencv-python
torch
torchvision
pyserial
pandas
requests
ultralytics